+++ src/Testing/TestExceptionHandler.php
@@ -50,6 +50,13 @@ class TestExceptionHandler extends ExceptionHandler
     use HandlesErrors;
 
     /**
+     * Whether Exceptions should be thrown to PHPUnit to handle.
+     *
+     * @var bool
+     */
+    public $rethrow = true;
+
+    /**
      * @var array
      * @todo when dropping support for Laravel 5.4, will no longer need to list these framework classes.
      */
@@ -69,7 +76,7 @@ class TestExceptionHandler extends ExceptionHandler
      */
     public function report(Exception $e)
     {
-        if ($this->shouldReport($e)) {
+        if ($this->shouldReport($e) && $this->rethrow) {
             throw $e;
         }
     }